slightly changed color of Informatik area of studies
This was necessary otherwise editing and saving the Methoden or Informatik area of studies without changing the colors would result in a color already existing error.
The reason for that is that these two demo data don't satisfy the minimum color offset, since the used setter method doesn't validate its input.
Validation only occurs during the post event of the area of studies create page.